requirements.txt
google-play-scraper>=1.2.2
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
python-dateutil>=2.8.0
tqdm>=4.65.0
python-dotenv>=1.0.0
scipy>=1.10.0
openpyxl>=3.1.0


